Various improvements for vector types, rgb, displayable

+++ src/Kalk.Core/Modules/Vectors/KalkColorRgbConstructor.cs
@@ -6,6 +6,6 @@
         {
         }
 
-        protected override KalkVector<int> NewVector(int dimension) => new KalkColorRgb();
+        protected override KalkVector<byte> NewVector(int dimension) => new KalkColorRgb();
     }
 }
